backend/requirements.txt
alembic==1.16.4
bcrypt==4.3.0
fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-multipart==0.0.20
SQLAlchemy==2.0.42
starlette==0.47.2
uvicorn==0.35.0
psycopg2==2.9.10
dnspython==2.7.0
email-validator==2.2.0